Add files via upload
Taking in a set of coordinates, This program tries to find a polynomial that defines a definite function f, behaving as the relation between the x value and the y value through f(x). The way undertaken in this program for that is the method of finite differences on polynomials as described here: https://en.wikipedia.org/wiki/Finite_difference, and here: https://brilliant.org/wiki/method-of-differences.If the program succeeds, the user can enter the word
 'more' to see the full solution.
written near-completely with vanilla python.